requirements.txt
Django==1.11
Pillow==5.0.0
boto==2.48.0
boto3==1.6.9
botocore==1.9.9
dj-database-url==0.5.0
django-forms-bootstrap==3.1.0
django-gravatar2==1.4.2
django-multiselectfield==0.1.8
django-storages==1.6.5
docutils==0.14
gunicorn==19.7.1
jmespath==0.9.3
psycopg2==2.7.4
python-dateutil==2.6.1
pytz==2018.3
s3transfer==0.1.13
stripe==1.79.1
